refactor: extract `<NuxtLayout>` from `<NuxtPage>`
