+++ docs/content/3.docs/2.directory-structure/15.app.md
@@ -25,7 +25,9 @@ If you have a [`pages/`](/docs/directory-structure/pages) directory, to display 
 ```vue [app.vue]
 <template>
   <div>
-    <NuxtPage/>
+    <NuxtLayout>
+      <NuxtPage/>
+    </NuxtLayout>
   </div>
 </template>
 ```
