+++ packages/nuxt3/src/pages/runtime/layout.ts
@@ -1,24 +1,33 @@
-import { defineComponent, h, Ref } from 'vue'
+import { defineComponent, isRef, Ref, Transition } from 'vue'
+import { useRoute } from 'vue-router'
+import { wrapIf } from './utils'
 // @ts-ignore
 import layouts from '#build/layouts'
+
+const defaultLayoutTransition = { name: 'layout', mode: 'out-in' }
 
 export default defineComponent({
   props: {
     name: {
       type: [String, Boolean, Object] as unknown as () => string | false | Ref<string | false>,
-      default: 'default'
+      default: null
     }
   },
   setup (props, context) {
+    const route = useRoute()
+
     return () => {
-      const layout = (props.name && typeof props.name === 'object' ? props.name.value : props.name) ?? 'default'
-      if (!layouts[layout]) {
-        if (process.dev && layout && layout !== 'default') {
-          console.warn(`Invalid layout \`${layout}\` selected.`)
-        }
-        return context.slots.default()
+      const layout = (isRef(props.name) ? props.name.value : props.name) ?? route.meta.layout as string ?? 'default'
+
+      const hasLayout = layout && layout in layouts
+      if (process.dev && layout && !hasLayout && layout !== 'default') {
+        console.warn(`Invalid layout \`${layout}\` selected.`)
       }
-      return h(layouts[layout], props, context.slots)
+
+      // We avoid rendering layout transition if there is no layout to render
+      return wrapIf(Transition, hasLayout && (route.meta.layoutTransition ?? defaultLayoutTransition),
+        wrapIf(layouts[layout], hasLayout, context.slots)
+      ).default()
     }
   }
 })
